// rh-lis/admin/src/app/models/interimaire.model.ts
export default class Interimaire {
  key?: string | null;
 cni?: string ;
 datenaiss?: string;
datesignature?: string;
  debutcontrat?: string;
 dpt?: string;
 fincontrat?: string;
 formation?: string;
  lieu_aff?: string;
  lieuaff?: string;
 lieunaiss?: string;
 lieusignature?: string;
 matricule?: string ;
nationnalite?: string;
 nom?: string;
 num?: string;
 poste?: string;
retenu?: string;
 salaire?: string;
 salairenet?: string;
tel_1?: string;
tel_2?: string;
type_pay?: string;
typecontrat?: string;
urgence1?: string;
urgence2?: string;
url?: string;
}